fix: signed cast for MD5 precompute chain-len guard; doc break; bump v1.5.8
Apply the explicit signed-cast pattern already used in precompute_ntlm8 to
precompute_md5_8/md5_9 across CUDA, Metal, and CL, so target_chain_len is
computed in signed arithmetic instead of relying on an implementation-defined
unsigned-to-signed conversion for the tail-thread guard.

Document the reduction_offset/chain_len compatibility break in CLAUDE.md so
pre-fix tables at nonzero table_index (or non-default MD5 chain lengths) are
flagged for regeneration.

Bump VERSION to v1.5.8.

+++ CL/precompute_md5_8.cl
@@ -20,7 +20,7 @@ __kernel void precompute_md5_8(
 
   /* Honor the host's chain_len (arg 8) instead of a hardcoded constant, so
    * lookups against tables of any chain length crack correctly. */
-  unsigned long chain_len = *g_chain_len;
+  long chain_len = (long)(*g_chain_len);
   long target_chain_len = (chain_len - *g_device_num) - ((get_global_id(0) + *g_exec_block_scaler) * *g_total_devices) - 1;
 
   if (target_chain_len < 1) {
