+++ Metal/precompute_md5_9.metal
@@ -24,7 +24,7 @@ kernel void precompute_md5_9(
 
   /* Honor the host's chain_len (arg 8) instead of a hardcoded constant, so
    * lookups against tables of any chain length crack correctly. */
-  ulong chain_len = *g_chain_len;
+  long chain_len = (long)(*g_chain_len);
   long target_chain_len = (chain_len - *g_device_num) - ((gid + *g_exec_block_scaler) * *g_total_devices) - 1;
 
   if (target_chain_len < 1) {
